5.3 beta 1 (#3335)
* Fix redirect checker for URLs with encoded characters

* Update packages

* Update the database install and upgrade component

Move away from classes and to function components. Include the importer step properly

* Typescript

* Clear file upload before leaving page

Stops certain edge case problems

* Prevent ordering by HTTP code column

* Code cleanup

* Encode URLs in nginx file

* Update composer modules

* Improve permalink migration

Try and handle the more difficult permalinks

* Default to no IP logging

* Remove precss

It’s causing lots of console warnings and doesn’t seem to be used

* Use WP CLI to generate pot file

* Move from Calypso i18n to @wordpress/i18n

* Use Gutenberg eslint

* Update language files

* Upgrade to React 18

* v1.2.2

* 5.3 beta 1

+++ database/database-status.php
@@ -73,7 +73,7 @@ class Red_Database_Status {
 		}
 
 		// Also if we're still in the process of upgrading
-		if ( $this->get_current_stage() ) {
+		if ( $this->get_current_stage() && $this->status !== self::STATUS_NEED_INSTALL ) {
 			return true;
 		}
 
@@ -88,16 +88,20 @@ class Red_Database_Status {
 	public function get_current_version() {
 		$settings = red_get_options();
 
-		if ( $settings['database'] !== '' ) {
+		if ( $settings['database'] !== '' && is_string( $settings['database'] ) ) {
 			return $settings['database'];
-		} elseif ( $this->get_old_version() !== false ) {
+		}
+
+		if ( $this->get_old_version() !== false ) {
 			$version = $this->get_old_version();
 
 			// Upgrade the old value
-			red_set_options( array( 'database' => $version ) );
-			delete_option( self::OLD_DB_VERSION );
-			$this->clear_cache();
-			return $version;
+			if ( $version ) {
+				red_set_options( array( 'database' => $version ) );
+				delete_option( self::OLD_DB_VERSION );
+				$this->clear_cache();
+				return $version;
+			}
 		}
 
 		return '';
